Testing Possibly Complete: getHumanPlayer() getComputerPlayer()
1-realized that by creating a game, then creating players, then starting
a new game that i was making two instance of a human and computer
(original problem when testing)

2-by looking at Game constructer, i realized that i could pull the
necessary qualities out from a freshly created game that could identify
the characteristics of a player (name and game)

Testing ran clean, but may want to look back to make sure about the
validity of this kind of test. Check with professor.