Developed a "How To" option on our JMenuBar
Now able to have the application link directly to the webpage that was
originally created for it

1-DON'T fill in the rest of the how to part of the web page until the
rest of the game revisions are made!

2-Implement the standard Mancala rule that says if the last stone you
place is in an empty pit on your side of the board, you will capture
(move to your store) any stones in the corresponding pit on the opposite
side of the board. Be sure to include some kind of notification that
lets the user know what just happened.

3-Implement the standard Mancala rules that says if the last stone you
place is in your store, you get to go again. Be sure to include some
kind of notification that lets the user know what just happened.

4-Note: website has been adjusted to avoid confusion. Nav panel changed
"How To" to "Original Rules." App may not have same rules as original
game in the end.