Game will allow user to play again if they place last seed in own store
•	(GAME CLASS) distributeStonesFromPit() = refactor to return int value
of last pit played
•	 (PLAYER INTERFACE) Player and all children = refactor the takeTurn()
method signature so that it returns an int value of the last pit played
•	(GAME CLASS) play() = asks us to have the currentPlayer takeTurn().
Store the last pit played when the current player takeTurn(). If
computer and last pit played is in computer’s store, then we will let
the computer play again (notify user). If human and last pit played is
human’s store, then we will let the human play again (notify user).