Finished updates to allow user to pick seeds per pit at start of game
To develop this part of the program, I first refactored two methods. 

First, I refactored the Game class’ resetBoard() method to accept an
integer value as a paremeter. This will allow me to reset the board
using any value between 1 and 4 (inclusive) that I want. 

Next, I refactored the Game class’ startnewGame() method to accept an
integer value to be used for the resetBoard() method. This allows me to
set up a new board via either method. 

From here, I went to the Gui class and altered the two action listeners
that check to see if it is the human or the computer who will start the
game. After browsing Oracle documentation, I was able to find a
JOptionPane that returns user feedback based on what they select in a
drop down menu. This feedback is what allows the user to determine how
many seeds should be in each pit. There is a cancel option that comes
with this pane by default, and will return null. As a default, I checked
for this and have the game start with a one seed per 
pet when this happens.